requirements.txt
pandas
chromadb
sentence-transformers
uuid
langchain
langchain-groq
streamlit
python-dotenv
groq
langdetect 
deep-translator
